refactor: new directory for component
