+++ src/Catalog.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import ProductCard from './ProductCard';
+import ProductCard from './components/ProductCard';
 
 import products from '../constants/Products';
